Fixed the thread pool and fixed an error in referencing the Empty exception.
