+++ billiard/pool.py
@@ -54,14 +54,12 @@ if platform.system() == 'Windows':  # pragma: no cover
     # *and its children* (if any).
     from ._win import kill_processtree as _kill  # noqa
 else:
-    from os import kill as _kill                 # noqa
-
+    from os import kill as _kill  # noqa
 
 try:
     TIMEOUT_MAX = threading.TIMEOUT_MAX
 except AttributeError:  # pragma: no cover
     TIMEOUT_MAX = 1e10  # noqa
-
 
 if sys.version_info >= (3, 3):
     _Semaphore = threading.Semaphore
@@ -201,6 +199,7 @@ class LaxBoundedSemaphore(_Semaphore):
             while self._Semaphore__value < self._initial_value:
                 _Semaphore.release(self)
 
+
 #
 # Exceptions
 #
@@ -230,12 +229,13 @@ class WorkersJoined(Exception):
 def soft_timeout_sighandler(signum, frame):
     raise SoftTimeLimitExceeded()
 
+
 #
 # Code run by worker processes
 #
 
 
-class Worker(Process):
+class WorkerMixin(object):
     _controlled_termination = False
     _job_terminated = False
 
@@ -252,7 +252,7 @@ class Worker(Process):
         self.inq, self.outq, self.synq = inq, outq, synq
         self._make_shortcuts()
 
-        super(Worker, self).__init__()
+        super(WorkerMixin, self).__init__(args=initargs)
 
     def __reduce__(self):
         return self.__class__, (
@@ -261,17 +261,23 @@ class Worker(Process):
         )
 
     def _make_shortcuts(self):
-        self.inqW_fd = self.inq._writer.fileno()    # inqueue write fd
-        self.outqR_fd = self.outq._reader.fileno()  # outqueue read fd
+        inq_has_writer = hasattr(self.inq, '_writer')
+        outq_has_writier = hasattr(self.outq, '_writer')
+
+        if inq_has_writer:
+            self.inqW_fd = self.inq._writer.fileno()  # inqueue write fd
+            self._quick_put = self.inq._writer.send
+        if outq_has_writier:
+            self.outqR_fd = self.outq._reader.fileno()  # outqueue read fd
+            self._quick_get = self.outq._reader.recv
+            self.send_job_offset = _get_send_offset(self.inq._writer)
+
         if self.synq:
             self.synqR_fd = self.synq._reader.fileno()  # synqueue read fd
             self.synqW_fd = self.synq._writer.fileno()  # synqueue write fd
             self.send_syn_offset = _get_send_offset(self.synq._writer)
         else:
             self.synqR_fd = self.synqW_fd = self._send_syn_offset = None
-        self._quick_put = self.inq._writer.send
-        self._quick_get = self.outq._reader.recv
-        self.send_job_offset = _get_send_offset(self.inq._writer)
 
     def run(self):
         _exit = sys.exit
@@ -280,6 +286,7 @@ class Worker(Process):
         def exit(status=None):
             _exitcode[0] = status
             return _exit()
+
         sys.exit = exit
 
         pid = os.getpid()
@@ -324,8 +331,8 @@ class Worker(Process):
     def workloop(self, debug=debug, now=monotonic, pid=None):
         pid = pid or os.getpid()
         put = self.outq.put
-        inqW_fd = self.inqW_fd
-        synqW_fd = self.synqW_fd
+        inqW_fd = self.inqW_fd if hasattr(self, 'inqW_fd') else None
+        synqW_fd = self.synqW_fd if hasattr(self, 'synqW_fd') else None
         maxtasks = self.maxtasks
         prepare_result = self.prepare_result
 
@@ -374,7 +381,7 @@ class Worker(Process):
                         ))
                         put((READY, (job, i, (False, einfo), inqW_fd)))
                     finally:
-                        del(tb)
+                        del (tb)
                 completed += 1
         debug('worker exiting after %d tasks', completed)
         if maxtasks:
@@ -392,16 +399,17 @@ class Worker(Process):
 
         # Make sure all exiting signals call finally: blocks.
         # This is important for the semaphore to be released.
-        reset_signals(full=self.sigprotection)
+        if threading.current_thread().__class__.__name__ == '_MainThread':
+            reset_signals(full=self.sigprotection)
 
-        # install signal handler for soft timeouts.
-        if SIG_SOFT_TIMEOUT is not None:
-            signal.signal(SIG_SOFT_TIMEOUT, soft_timeout_sighandler)
+            # install signal handler for soft timeouts.
+            if SIG_SOFT_TIMEOUT is not None:
+                signal.signal(SIG_SOFT_TIMEOUT, soft_timeout_sighandler)
 
-        try:
-            signal.signal(signal.SIGINT, signal.SIG_IGN)
-        except AttributeError:
-            pass
+            try:
+                signal.signal(signal.SIGINT, signal.SIG_IGN)
+            except AttributeError:
+                pass
 
     def _make_recv_method(self, conn):
         get = conn.get
@@ -422,7 +430,7 @@ class Worker(Process):
             def _recv(timeout):  # noqa
                 try:
                     return True, get(timeout=timeout)
-                except Queue.Empty:
+                except Empty:
                     return False, None
         return _recv
 
@@ -456,13 +464,22 @@ class Worker(Process):
         return receive
 
 
+class Worker(WorkerMixin, Process):
+    pass
+
+
+class ThreadWorker(WorkerMixin, DummyProcess):
+    @property
+    def pid(self):
+        return self.ident
+
+
 #
 # Class representing a process pool
 #
 
 
 class PoolThread(DummyProcess):
-
     def __init__(self, *args, **kwargs):
         DummyProcess.__init__(self)
         self._state = RUN
@@ -503,7 +520,6 @@ class PoolThread(DummyProcess):
 
 
 class Supervisor(PoolThread):
-
     def __init__(self, pool):
         self.pool = pool
         super(Supervisor, self).__init__()
@@ -540,7 +556,6 @@ class Supervisor(PoolThread):
 
 
 class TaskHandler(PoolThread):
-
     def __init__(self, taskqueue, put, outqueue, pool):
         self.taskqueue = taskqueue
         self.put = put
@@ -602,7 +617,6 @@ class TaskHandler(PoolThread):
 
 
 class TimeoutHandler(PoolThread):
-
     def __init__(self, processes, cache, t_soft, t_hard):
         self.processes = processes
         self.cache = cache
@@ -613,9 +627,9 @@ class TimeoutHandler(PoolThread):
 
     def _process_by_pid(self, pid):
         return next((
-            (proc, i) for i, proc in enumerate(self.processes)
-            if proc.pid == pid
-        ), (None, None))
+                        (proc, i) for i, proc in enumerate(self.processes)
+                        if proc.pid == pid
+                    ), (None, None))
 
     def on_soft_timeout(self, job):
         debug('soft time limit exceeded for %r', job)
@@ -660,8 +674,9 @@ class TimeoutHandler(PoolThread):
         except OSError:
             pass
         else:
-            if worker._popen.wait(timeout=0.1):
-                return
+            if hasattr(worker, '_popen'):
+                if worker._popen.wait(timeout=0.1):
+                    return
         debug('timeout: TERM timed-out, now sending KILL to %s', worker._name)
         try:
             _kill(worker.pid, signal.SIGKILL)
@@ -722,7 +737,6 @@ class TimeoutHandler(PoolThread):
 
 
 class ResultHandler(PoolThread):
-
     def __init__(self, outqueue, get, cache, poll,
                  join_exited_workers, putlock, restart_state,
                  check_timeouts, on_job_ready):
@@ -790,6 +804,7 @@ class ResultHandler(PoolThread):
                 state_handlers[state](*args)
             except KeyError:
                 debug("Unknown job state: %s (args=%s)", state, args)
+
         self.on_state_change = on_state_change
 
     def _process_result(self, timeout=1.0):
@@ -950,6 +965,9 @@ class Pool(object):
             ))
             soft_timeout = None
 
+        if processes is not None and processes < 1:
+            raise ValueError("A pool must have at least one worker.")
+
         self._processes = self.cpu_count() if processes is None else processes
         self.max_restarts = max_restarts or round(self._processes * 100)
         self.restart_state = restart_state(max_restarts, max_restart_freq or 1)
@@ -1039,9 +1057,9 @@ class Pool(object):
 
     def _process_by_pid(self, pid):
         return next((
-            (proc, i) for i, proc in enumerate(self._pool)
-            if proc.pid == pid
-        ), (None, None))
+                        (proc, i) for i, proc in enumerate(self._pool)
+                        if proc.pid == pid
+                    ), (None, None))
 
     def get_process_queues(self):
         return self._inqueue, self._outqueue, None
@@ -1094,11 +1112,13 @@ class Pool(object):
         for i in reversed(range(len(self._pool))):
             worker = self._pool[i]
             exitcode = worker.exitcode
-            popen = worker._popen
+            popen = worker._popen if hasattr(worker, '_popen') else None
             if popen is None or exitcode is not None:
                 # worker exited
                 debug('Supervisor: cleaning up worker %d', i)
                 if popen is not None:
+                    worker.join()
+                elif isinstance(worker, ThreadWorker):
                     worker.join()
                 debug('Supervisor: worked %d joined', i)
                 cleaned[worker.pid] = worker
@@ -1267,6 +1287,7 @@ class Pool(object):
 
     def _setup_queues(self):
         from billiard.queues import SimpleQueue
+
         self._inqueue = SimpleQueue()
         self._outqueue = SimpleQueue()
         self._quick_put = self._inqueue._writer.send
@@ -1276,6 +1297,7 @@ class Pool(object):
             if self._outqueue._reader.poll(timeout):
                 return True, self._quick_get()
             return False, None
+
         self._poll_result = _poll_result
 
     def _start_timeout_handler(self):
@@ -1431,7 +1453,7 @@ class Pool(object):
                 self._start_timeout_handler()
             if self.threads:
                 self._taskqueue.put(([(TASK, (result._job, None,
-                                    func, args, kwds))], None))
+                                              func, args, kwds))], None))
             else:
                 self._quick_put((TASK, (result._job, None, func, args, kwds)))
             return result
@@ -1528,8 +1550,10 @@ class Pool(object):
         stop_if_not_current(self._result_handler)
         debug('result handler joined')
         for i, p in enumerate(self._pool):
-            debug('joining worker %s/%s (%r)', i+1, len(self._pool), p)
-            if p._popen is not None:  # process started?
+            debug('joining worker %s/%s (%r)', i + 1, len(self._pool), p)
+            if hasattr(p, '_popen') and p._popen is not None:  # process started?
+                p.join()
+            elif isinstance(p, DummyProcess):
                 p.join()
         debug('pool join complete')
 
@@ -1562,7 +1586,7 @@ class Pool(object):
         worker_handler.terminate()
 
         task_handler.terminate()
-        taskqueue.put(None)                 # sentinel
+        taskqueue.put(None)  # sentinel
 
         debug('helping task handler/workers to finish')
         cls._help_stuff_finish(*help_stuff_finish_args)
@@ -1596,13 +1620,16 @@ class Pool(object):
                 if p.is_alive():
                     # worker has not yet exited
                     debug('cleaning up worker %d', p.pid)
-                    if p._popen is not None:
+                    if hasattr(p, '_popen') and p._popen is not None:
+                        p.join()
+                    elif isinstance(p, DummyProcess):
                         p.join()
             debug('pool workers joined')
 
     @property
     def process_sentinels(self):
-        return [w._popen.sentinel for w in self._pool]
+        return [w._popen.sentinel for w in self._pool if hasattr(w, '_popen')]
+
 
 #
 # Class whose instances are returned by `Pool.apply_async()`
@@ -1761,13 +1788,13 @@ class ApplyResult(object):
             if self._send_ack and synqW_fd:
                 self._send_ack(response, pid, self._job, synqW_fd)
 
+
 #
 # Class whose instances are returned by `Pool.map_async()`
 #
 
 
 class MapResult(ApplyResult):
-
     def __init__(self, cache, chunksize, length, callback, error_callback):
         ApplyResult.__init__(
             self, cache, callback, error_callback=error_callback,
@@ -1822,6 +1849,7 @@ class MapResult(ApplyResult):
     def worker_pids(self):
         return [pid for pid in self._worker_pid if pid]
 
+
 #
 # Class whose instances are returned by `Pool.imap()`
 #
@@ -1868,7 +1896,7 @@ class IMapIterator(object):
             return value
         raise Exception(value)
 
-    __next__ = next                    # XXX
+    __next__ = next  # XXX
 
     def _set(self, i, obj):
         with self._cond:
@@ -1904,13 +1932,13 @@ class IMapIterator(object):
     def worker_pids(self):
         return self._worker_pids
 
+
 #
 # Class whose instances are returned by `Pool.imap_unordered()`
 #
 
 
 class IMapUnorderedIterator(IMapIterator):
-
     def _set(self, i, obj):
         with self._cond:
             self._items.append(obj)
@@ -1920,18 +1948,18 @@ class IMapUnorderedIterator(IMapIterator):
                 self._ready = True
                 del self._cache[self._job]
 
+
 #
 #
 #
 
 
 class ThreadPool(Pool):
-
     from billiard.dummy import Process as DummyProcess
+
     Process = DummyProcess
 
-    def __init__(self, processes=None, initializer=None, initargs=()):
-        Pool.__init__(self, processes, initializer, initargs)
+    Worker = ThreadWorker
 
     def _setup_queues(self):
         self._inqueue = Queue()
@@ -1944,6 +1972,7 @@ class ThreadPool(Pool):
                 return True, self._quick_get(timeout=timeout)
             except Empty:
                 return False, None
+
         self._poll_result = _poll_result
 
     @staticmethod
